Tags and Sub Verification (#4)
Update tags position based on bazarr labeling

Change Tags position based on bazarr naming

Cleaned up and add ability to use TYPE.LANG or LANG.TYPE (Bazarr format)

Cleanup logic

Updated to incorporate new function.

Test infinite loop

Multiple Sub Check Testing.  To confirm new subtitle was added.

Clean up and finalize testing

+++ mergesrt.sh
@@ -44,20 +44,64 @@ process() {
     # PARSE FILE COMPONENTS ------------------------------------------------------------
     EXT=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f1 | rev)
     echo -e "\e[1;34mExtension: $EXT\e[m"
-    if [ "$EXT" == "srt" ]; then
-        LANG=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev)
-        echo -e "\e[1;34mSubtitle language: $LANG\e[m"
-        TYPE=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev)
-        if [ "$TYPE" == 'sdh' ] || [ "$TYPE" == 'forced' ] || [ "$TYPE" == 'hi' ] || [ "$TYPE" == 'cc' ]; then
-            echo -e "\e[1;34mSubtitle type: $TYPE\e[m"
-            FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$TYPE"'\.'"$LANG"'\.'"$EXT"'||')
-        else 
-            TYPE=""
-            FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$LANG"'\.'"$EXT"'||')
-        fi
-    else
-        FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$EXT"'||')
-    fi
+  
+    case "$EXT" in
+        srt)
+            # Check file endings against a ISO 639-1 and ISO 639-2 list to determine if valid LANG
+            F2="$(curl -s -L https://datahub.io/core/language-codes/r/1.csv | grep -ow $(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev))"
+            F3="$(curl -s -L https://datahub.io/core/language-codes/r/1.csv | grep -ow $(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev))"
+
+            # Check Language
+            if [ -n "$F2" ] && [ -z "$F3" ]; then # Check if first valid after .EXT
+                LANG=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev)
+                TYPE=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev)
+                TYPE_LOC=2
+            elif [ -n "$F3" ] && [ -z "$F2" ]; then # Check if second valid after .EXT
+                LANG=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev)
+                TYPE=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev)
+                TYPE_LOC=1
+            elif [ -n "$F2" ] && [ -n "$F3" ]; then # Check if both valid (edge case of using hi)
+                # Use hi as TYPE (should help prevent Hindu 2 lang code issues)
+                if [ $(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev) == 'hi' ]; then
+                    LANG=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev)
+                    TYPE=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev)
+                    TYPE_LOC=1
+                else
+                    LANG=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f2 | rev)
+                    TYPE=$(echo "$IMPORT_FILE" | rev | cut -d'.' -f3 | rev)
+                    TYPE_LOC=2
+                fi
+            else
+                echo -e "\e[0;31mCould not determine file language, skipping\e[m"
+                return
+            fi
+            echo -e "\e[1;34mSubtitle language: $LANG\e[m"
+
+            #Check TYPE
+            if [ "$TYPE" == 'sdh' ] || [ "$TYPE" == 'forced' ] || [ "$TYPE" == 'hi' ] || [ "$TYPE" == 'cc' ]; then
+                echo -e "\e[1;34mSubtitle type: $TYPE\e[m"
+            else 
+                TYPE_LOC=0
+            fi
+
+            # Generate Correct File Name
+            case $TYPE_LOC in
+            0)
+                FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$LANG"'\.'"$EXT"'||')
+                ;;
+            1)
+                FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$LANG"'\.'"$TYPE"'\.'"$EXT"'||')
+                ;;
+            2)
+                FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$TYPE"'\.'"$LANG"'\.'"$EXT"'||')
+                ;;
+            esac
+            ;;
+        idx)
+            FILE_NAME=$(echo "$IMPORT_FILE" | sed 's|\.'"$EXT"'||')
+            ;;
+    esac
+    
     echo -e "\e[1;34mFile name: $FILE_NAME\e[m"
     
     VIDEO_FILE=$FILE_NAME'.mkv'
@@ -70,14 +114,18 @@ process() {
         return
     fi
     echo -e "\e[1;32mSTARTING MERGE\e[m"
-    MERGE_FILE=$FILE_NAME'.merge'
+    MERGE_FILE=$FILE_NAME'.merge'    
+    CURR_SUB_COUNT="$(mkvmerge --identify "$VIDEO_FILE" | grep -c 'subtitle')" # Count the number of subs in the pre processed file
     merge "$MERGE_FILE" "$VIDEO_FILE" "$IMPORT_FILE" "$EXT" "$TYPE" "$LANG"
     # When doing large batches sometimes the merge does not seem to work correctly.
-    # this is used to keep running the merge untill the file has detected a subtitle.
-    while !(mkvmerge --identify "$MERGE_FILE" | grep -c -q 'subtitle') do
+    # this is used to keep running the merge until the file has detected a new subtitle.
+    NEXT_SUB_COUNT="$(mkvmerge --identify "$MERGE_FILE" | grep -c 'subtitle')" # Count the number of subs in the post processed file
+    while [ "$CURR_SUB_COUNT" -gt "$NEXT_SUB_COUNT" ] 
+    do
         echo -e "\e[0;31mSubtitle is missing from merge file.  Rerunning merge\e[m"
         rm "$MERGE_FILE"
         merge "$MERGE_FILE" "$VIDEO_FILE" "$IMPORT_FILE" "$EXT" "$TYPE" "$LANG"
+        NEXT_SUB_COUNT="$(mkvmerge --identify "$MERGE_FILE" | grep -c 'subtitle')"
     done
     RESULT=$?
     # CLEAN UP  --------------------------------------------------------------------------
@@ -86,12 +134,13 @@ process() {
         echo "$RESULT"
         echo "Delete $IMPORT_FILE"
         rm "$IMPORT_FILE"
-        rm "$FILE_NAME.sub"
+        if ("$EXT" == 'idx'); then
+            rm "$FILE_NAME.sub"
+        fi
         echo "Delete $VIDEO_FILE"
         rm "$VIDEO_FILE"
         echo "Rename $MERGE_FILE to $FILE_NAME.mkv"
         mv "$MERGE_FILE" "$FILE_NAME.mkv"
-        # rm "$MERGE_FILE"
         echo "---------------------------- END PROCESS ---------------------------"
     else
         echo -e "\e[0;31mMERGE FAILED\e[m"
